@@ -30,7 +30,7 @@
 package org.moe.samples.openglbox.android;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import org.moe.samples.openglbox.android.R;
 
